+++ src/lib/fh.h
@@ -11,7 +11,7 @@ struct fh_input;
 struct fh_program;
 
 struct fh_input_funcs {
-  ssize_t (*read)(struct fh_input *in, char *line, ssize_t max_len);
+  int (*read)(struct fh_input *in, char *line, int max_len);
   int (*close)(struct fh_input *in);
 };
 
@@ -45,7 +45,7 @@ struct fh_input *fh_open_input_file(const char *filename);
 struct fh_input *fh_new_input(void *user_data, struct fh_input_funcs *funcs);
 void *fh_get_input_user_data(struct fh_input *in);
 int fh_close_input(struct fh_input *in);
-ssize_t fh_input_read(struct fh_input *in, char *line, ssize_t max_len);
+int fh_input_read(struct fh_input *in, char *line, int max_len);
 
 struct fh_program *fh_new_program(void);
 void fh_free_program(struct fh_program *prog);
